Remove recipe for installing Kafka from source
With the 0.8.2 release there’ll be three different ways to build three
different versions of Kafka, and this is just getting out of hand.
There’s too much to support between releases, and I highly doubt that
anyone is actually gonna deploy Kafka from source to production servers.

The `binary` recipe has been renamed to `_install` and is considered an
“internal” recipe, other than that there’s just a lot of code gone, and
there’s a lot more that can be cleaned up.